pyproject.toml
[project]
name = "llm-mcp-server"
version = "0.1.0"
description = "MCP server for querying Claude and ChatGPT"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
